refactor(backserver): 合并 back_server 两文件 + 统一命名（#737 收尾）
- 合并：back_server_inner.go（startStaticServer/cors/setUpRouters）并入
  back_server.go，删除 back_server_inner.go，消除「物理拆分 vs 职责」sharp-edge。
- 改名（grep 量化后机械重命名）：
  - BackServer.DictCon → Controller（field + 访问，app.go/backserver）。
  - model.DictionaryItem.MainDict → Dict（运行时字段，非 JSON 序列化；dict_def/
    dict_item/dicts_service/dicts_handler + stardict 测试）。
  - DictsController.ds 字段 → svc（field/构造器/参数 + dc.ds 访问）。
  注：DictService 方法接收者 ds 保留（Go 短接收者约定）。
- CLAUDE.md：更新 back_server split sharp-edge（已合并 + IPC 已 typed）。

go build/test(-race)/vet 全绿。

Closes #737
