+++ pkg/backserver/back_server.go
@@ -18,17 +18,20 @@ package backserver
 
 import (
 	"context"
-	"github.com/terasum/medict/pkg/apis"
 	"net"
 	"net/http"
 	"strings"
 	"time"
 
 	"github.com/gin-gonic/gin"
+	"github.com/op/go-logging"
 	"github.com/terasum/medict/internal/config"
 	"github.com/terasum/medict/internal/static"
+	"github.com/terasum/medict/pkg/apis"
 	"github.com/terasum/medict/pkg/service"
 )
+
+var log = logging.MustGetLogger("backserver")
 
 type BackServer struct {
 	Config *config.Config
@@ -43,8 +46,8 @@ type BackServer struct {
 	Ready   bool
 	DevMode bool
 
-	GinEngine *gin.Engine
-	DictCon   *apis.DictsController
+	GinEngine  *gin.Engine
+	Controller *apis.DictsController
 }
 
 func NewStaticServer(conf *config.Config) (*BackServer, error) {
@@ -58,9 +61,9 @@ func NewStaticServer(conf *config.Config) (*BackServer, error) {
 }
 
 // SetUp wires the injected DictService into the controller and registers
-// handlers/routes. The DictService is owned by App and passed in (issue #727).
+// routes. The DictService is owned by App and passed in (issue #727).
 func (bs *BackServer) SetUp(dictsSvc *service.DictService) error {
-	bs.DictCon = apis.NewDictsController(dictsSvc)
+	bs.Controller = apis.NewDictsController(dictsSvc)
 
 	if err := bs.setUpRouters(); err != nil {
 		return err
@@ -108,4 +111,109 @@ func (bs *BackServer) StaticServerBaseUrl() string {
 		return listenAddr
 	}
 	return ""
+}
+
+func (bs *BackServer) startStaticServer(listenAddr string) {
+	if listenAddr == "" {
+		listenAddr = "localhost:0"
+	}
+	if listenAddr == ":0" {
+		listenAddr = "localhost:0"
+	}
+
+	srv := &http.Server{
+		Addr:    listenAddr, // use next port available
+		Handler: bs.GinEngine,
+	}
+
+	log.Infof("start listening... %s\n", srv.Addr)
+
+	ln, err := net.Listen("tcp", listenAddr)
+	if err != nil && err != http.ErrServerClosed {
+		log.Infof("backserver listen err: %s\n", err)
+	} else {
+		log.Infof("backserver listen at: %s\n", ln.Addr().String())
+	}
+
+	bs.ListenAddr = ln.Addr()
+
+	go func() {
+		bs.Ready = true
+		err = srv.Serve(ln)
+		if err != nil && err != http.ErrServerClosed {
+			log.Infof("backserver serve err: %s\n", err)
+		}
+	}()
+
+	bs.Srv = srv
+}
+
+// allowedOrigin reports whether the given Origin is permitted to make
+// cross-origin requests to the embedded resource server.
+//
+// 正式功能不依赖跨域：前端控制面走 Wails IPC，释义 HTML 及其子资源经同源
+// iframe（http://localhost:<port>）加载，本身不需要 CORS。此处仅放行本地
+// wails webview 与 loopback 调试来源，避免原实现反射任意 Origin 的安全隐患。
+func allowedOrigin(origin string) bool {
+	if origin == "" {
+		return false
+	}
+	switch origin {
+	case "http://wails.localhost", "https://wails.localhost",
+		"wails://wails.localhost", "wails://localhost":
+		return true
+	}
+	for _, scheme := range []string{"http", "https"} {
+		for _, host := range []string{"localhost", "127.0.0.1"} {
+			if origin == scheme+"://"+host || strings.HasPrefix(origin, scheme+"://"+host+":") {
+				return true
+			}
+		}
+	}
+	return false
+}
+
+func cors() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		method := c.Request.Method
+		origin := c.Request.Header.Get("Origin")
+		if allowedOrigin(origin) {
+			c.Writer.Header().Set("Access-Control-Allow-Origin", origin)
+			c.Header("Access-Control-Allow-Methods", "POST, GET, OPTIONS, PUT, DELETE,UPDATE")
+			c.Header("Access-Control-Allow-Headers", "Authorization, Content-Length, X-CSRF-Token, Token,session")
+			c.Header("Access-Control-Expose-Headers", "Content-Length, Access-Control-Allow-Origin, Access-Control-Allow-Headers")
+			c.Header("Access-Control-Max-Age", "172800")
+			c.Header("Access-Control-Allow-Credentials", "true")
+		}
+
+		if method == "OPTIONS" {
+			c.JSON(http.StatusOK, "ok")
+		}
+
+		defer func() {
+			if err := recover(); err != nil {
+				log.Infof("Panic info is: %v\n", err)
+			}
+		}()
+
+		c.Next()
+	}
+}
+
+func (bs *BackServer) setUpRouters() error {
+	bs.GinEngine.Use(cors())
+
+	// Word lookup: an explicit route (previously a string-HasPrefix branch
+	// inside NoRoute — issue #728). Discoverable, middleware-able, structured
+	// errors via the handler instead of catch-all dispatch.
+	bs.GinEngine.GET(static.ContentRootUrl+static.WordQueryMagicPath, bs.Controller.HandleWordQueryReq)
+
+	// Everything else is a resource lookup. Resource paths are arbitrary
+	// (css / images / fonts / ... under the content root), so they stay a
+	// catch-all rather than enumerated routes.
+	bs.GinEngine.NoRoute(func(c *gin.Context) {
+		log.Debugf("resource request: %s", c.Request.RequestURI)
+		bs.Controller.HandleResourceQueryReq(c)
+	})
+	return nil
 }
